INTERNAL [IOTG] drm/i915: Introduce initial mode setting
For very specific use cases it may be desirable to have the kernel
driver do an initial mode set and/or splash screen type display without
having to enable a framebuffer console.

A primary goal for this feature is to have the display lit up as soon
as possible.  A non-goal is the ability to detect and adapt to the
display configurations. In other words, the  display configuration will
be a known/fixed configuration.  In addition, it is expected that the
EDID information will be present and optimized so mode list creation
and traversal time is minimized.

Add a new module parameter that replaces the normal fbdev initialization
with a new initial mode set.  When the parameter is off, follow the normal
program flow and (if configured) initialize fbdev and a framebuffer
console.  When the parameter is on, do an atomic mode set with everything
off (no crtc's enabled).

Signed-off-by: Bob Paauwe <bob.j.paauwe@intel.com>
Reviewed-by: Matt Roper <matthew.d.roper@intel.com>
Reviewed-by: James Xiong <james.xiong@intel.com>
Signed-off-by: Matt Roper <matthew.d.roper@intel.com>
Change-Id: I261902f14be9aabd060ae164f055e881fecf17ff