ASoC: Intel: Skylake: Probe DMA release for extractor
1. Extractor DMA is to be assigned when the first probe stream (irrespective
of whether it is injector or extractor) is opened. But if the first probe
stream is injector, we get injector's substream pointer and we do not
have the right substream pointer for extractor.

The existing code passed injector's substream while assigning extractor
DMA.This patch sets NULL for substream while assigning the DMA for
extractor and sets the correct substream pointer later when open is
indeed for extractor.

2. DMA reset for extractor should not be done after  probe module is
initialized.
The existing code reset DMA in hw_params. This can result in DMA reset
after probe module init when  injector and extractor are started one
after the other.

3. Extractor DMA is assigned in compr_open for the first probe stream
irrespective of whether it is injector or extractor.
So DMA release for extractor should be done in the case where
a injector probe alone was started and stopped without starting
any extractor.

This patch moves the DMA reset from hw params to immediately after DMA
assignment in open call back for both injector and extractor.

Change-Id: I2604796d81e2e6da5acd3977774887a0b2e14559
Signed-off-by: Pawse, GuruprasadX <guruprasadx.pawse@intel.com>
Signed-off-by: Mohit Sinha <mohit.sinha@intel.com>
Reviewed-by: Shaik, Kareem M <kareem.m.shaik@intel.com>
Reviewed-by: audio_build <audio_build@intel.com>
Reviewed-by: Babu, Ramesh <ramesh.babu@intel.com>
Tested-by: Avati, Santosh Kumar <santosh.kumar.avati@intel.com>