rpmb: enable emmc specific read data fixup
For eMMC the block count of the RPMB read operation is not indicated in
the original RPMB Data Read Request packet.
This might be different then the implementation of other protocol
standards.
This patch implements a fixup for this behavior.

V6: New in the series.
V7: Resend
V8: Resend.
V9: Scan all the frames in the sequence.

Change-Id: I7292ce1e6f89e4a77e61a3b9208704bf4add05c4
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Tested-by: Avri Altman <avri.altman@sandisk.com>