MUST_REBASE [VPG]: drm/i915: Allow late GuC/HuC loading
This patch allows for the GuC and HuC to be loaded until the first
device is open (instead of during the driver initialisation).

Delaying the fw load prevents a load failure when the partition that the
firmware is supposed to be on has not been mounted yet, as is the case
in Android.

Since the GuC won't exist until later on, a couple of functions with
need of it have to also be delayed (i915_guc_log_register &
__intel_autoenable_gt_powersave).

There was a collection of patches in 4.1/4.4 to do more less the same,
but they were obsolete for 4.11 and it made more sense to rewrite it
from scratch.

MUST_REBASE: The hw_late code is part of a solution for loading firmware
of a file system that has not been mounted yet. Only a problem
on Android.

Change-Id: I8bc36eca473dcaded3f52edafb65d315f949083c
Signed-off-by: Michel Thierry <michel.thierry@intel.com>