extcon: USB: Use static DRD mode for host/device mode switch
Currently, dynamic DRD mode is used for handling the host/device mode
switch. But this method creates issues when switching the mode after S3
or after coldboot. This issue can be avoided if we use static DRD mode
for the host/device mode switch. This patch implements this support.

In addition, the ABL initial the DRD as dynamic, for keep ABL
compatiable for both Yocto and Android, we initial the DRD mode in kernel
and keep no change in ABL.

Change-Id: Ib67884e57a7545a367c98371a417ca5c2099c541
Signed-off-by: he, bo <bo.he@intel.com>
Signed-off-by: Anuar, Nuhairi <nuhairi.anuar@intel.com>
Signed-off-by: Kuppuswamy, Sathyanarayanan <sathyanarayanan.kuppuswamy@intel.com>