ASoC: Intel: Topology: Add dynamic module id support
Module id is a property of fw manifest and can vary between platforms.
So use the uuid instead of module id for pins.

With this patch, skl_tplg_get_uuid() is rearranged, so that it can be
used during skl_tplg_fill_pin() and arguments to skl_tplg_fill_pin() are
modified to take the snd_soc_tplg_vendor_value_elem as argument so that
both value and uuid can be referred to to fill the pin information.

Change-Id: Ic7ca9c367fa63d5eacc5476fef7ede3c3c975446
Signed-off-by: Periyasamy, SriramX <sriramx.periyasamy@intel.com>
Reviewed-by: Prusty, Subhransu S <subhransu.s.prusty@intel.com>
Reviewed-by: Sm, Bhadur A <bhadur.a.sm@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>