ASoC: Intel: Skylake: Support multiple format configs
A module can have two kinds of set params, as per
topology requirements. For example, one pre-init and
one post-init. But currently, there is support for
just one type, as the format_config.

This patch extends the format_configs to 4, so as to
be able to support pre-init, post-init and post-bind
type of set params, for the same module, simultaneously.

Change-Id: I5cdf6a921db6e9ffcabda2ec601795564c8f53d2
Signed-off-by: Kareem Shaik <kareem.m.shaik@intel.com>
Reviewed-by: Gogineni, GiribabuX <giribabux.gogineni@intel.com>
Reviewed-by: Kale, Sanyog R <sanyog.r.kale@intel.com>
Reviewed-by: Singh, Guneshwor O <guneshwor.o.singh@intel.com>
Reviewed-by: Nc, Shreyas <shreyas.nc@intel.com>
Reviewed-by: Kesapragada, Pardha Saradhi <pardha.saradhi.kesapragada@intel.com>
Reviewed-by: Koul, Vinod <vinod.koul@intel.com>
Tested-by: Sm, Bhadur A <bhadur.a.sm@intel.com>