ASoC: Intel: Skylake: Select BXT-P machine driver based on kernel config
The machine driver id is same for ULL and IVI, Current implementation of
loading the machine driver is based on machine id, it always look for
IVI machine driver as this the first entry in the table and
fails to load ULL machine driver.

Use configuration option to select correct machine driver defined under
same machine id.

Change-Id: I4d9115db1f5be598db311858dd003526ae97b767
Signed-off-by: Giribabu Gogineni <giribabux.gogineni@intel.com>