drm/i915: Repair the preemption context if hit by reset
It is possible for the preemption context to be active on an engine
when the engine is reset. The preemption context is not tracked via
the request mechanism, so the normal reset handling will not detect
this scenario and will not be able to fixup possible context
corruption. So add some extra logic to do this.

This patch is required to support the force preemption feature.
There are a series of patches upstream which insulate the preemption
context from corruption, making this repair unecessary. However it
is an extensive backport that modifies other functionality and is
a bit risky. So this simple point fix seems the better option.

Test: Run IGT gem_exec_fpreempt repeatedly.
Change-Id: I845670d85bf65347ce31d0f6a3479c3cf6e2effd
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>