media: intel-ipu4: imx319: Be compatible for kernel 4.4 and 4.14
Message for Open Source:
Wrap code with kernel version checking to make driver be compatible
with both kernel 4.4 and 4.14.

Message for Internal:
[Issue/Feature] imx319 driver needs kernel version adapation as it is
used for both KBL(kernel 4.4) and CNL/ICL(kernel 4.14)
[Root Cause/Changes] Wrap code to make driver compatible for both
kernel 4.4 and 4.14. v4l2_async_register_subdev_sensor_common()
is a special case. It has been upstreamed but does not exist in
kernel 4.4 and 4.14. Google kernel has back ported this patch
so it has this function. As this driver is only used in
google kernel, no macro wraping is applied for this function.

Change-Id: I37480cf4e97d2ff89d3be86f2b9b17f81de87702
Signed-off-by: Tianshu Qiu <tian.shu.qiu@intel.com>
Signed-off-by: Meng Wei <wei.meng@intel.com>