usb: roles: intel_xhci: Set device mode as default.
In APL SOC, if there is no USB device connected to xHCI port, then
we are not able to override the mux mode in CFG0 register. So we
cannot switch the role after booting the system using sysfs file.
So forcing the default mode to USB_DEVICE, till we root cause this
iproblem.

Signed-off-by: Kuppuswamy Sathyanarayanan <sathyanarayanan.kuppuswamy@linux.intel.com>