media: intel-ipu4: remove some include files
Message for Open Source:
The header file itself was just a collection of structure
definitions, clean up its headers to less dependency.

Message for Internal:
[Issue/Feature]
There is user space program which includes this file and it will
complain no header file found. If no harm to the kernel space, remove
the header include.
[Root Cause/Changes]

Change-Id: Ief6847e40a30a976bda3d522265fa45977ebd9f3
Signed-off-by: Ding Yunliang <yunliang.ding@intel.com>
Signed-off-by: Meng Wei <wei.meng@intel.com>