media: intel-ipu4: depracated abi sanity check
Message for Open Source:
The abi check is now having heavy maintain burden, and issue was not
easy to find, also makes code hard to read. So move this check
outside of kernel space, as this is an internal match check,
should not bother kernel users.

Message for Internal:
[Issue/Feature]
[Root Cause/Changes]
Move to icg-sw-devtools.

Change-Id: I471304d10a6dd92636448ea8f82f553044d27894
Signed-off-by: Ding Yunliang <yunliang.ding@intel.com>
Signed-off-by: Meng Wei <wei.meng@intel.com>