drm/i915: to limit the supported modifiers for plane restriction
In GVT-g environment, to ensure all the OS's have enough DDB to
display, GVT-g will statically allocate all the DDBs for all the
planes on all the pipes.
However, when SOS or UOS wants to use Y/Yf tiled modifier, the watermark
required will exceed the DDBs allocated by GVT-g, thus causes some
display issues.

So in this patch, we removed the supports of the Y/Yf tiled modifiers
for both SOS and UOS when plane restriction is enabled. And a consequence
is that the RBC will be disabled since _CCS modifiers will no longer be
supported.

Signed-off-by: Min He <min.he@intel.com>
Reviewed-by: Zhao Yakui <yakui.zhao@intel.com>
Reviewed-by: Fei Jiang <fei.jiang@intel.com>