Initial implementation to test the SWIOTLB
This driver is to be used for regression testing of SWIOTLB.
It allocates a fake device, allocates a page(s) and performs
a variety of DMA operations with seeding the DMA page and
the destination page (or vice-versa) with expected values.
After the DMA operations are done the values are checked.

Signed-off-by: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
[v1: Initial implementation copied from git@github.com:konradwilk/swiotlb-test.git]